Fixed stall endstop trigger time being truncated to 16 bits
The trigger time passed to HandleStalledRemoteDrivers was narrowed to uint16_t by the virtual it goes through.
RRF then evaluated the axis position at a bogus time and reverted the drivers back to almost the start of the homing move

+++ src/Endstops/StallDetectionEndstop.cpp
@@ -290,7 +290,7 @@ void StallDetectionEndstop::DeleteRemoteStallEndstops() noexcept
 }
 
 // Record any notifications of stalled remote drivers that we are interested in
-void StallDetectionEndstop::HandleStalledRemoteDrivers(CanAddress boardAddress, LocalDriversBitmap driversReportedStalled, uint16_t when) noexcept
+void StallDetectionEndstop::HandleStalledRemoteDrivers(CanAddress boardAddress, LocalDriversBitmap driversReportedStalled, uint32_t when) noexcept
 {
 	remoteDriversMonitored.IterateWhile([this, boardAddress, driversReportedStalled, when](RemoteDriversMonitored& entry, size_t count) noexcept -> bool
 										{
